coffe_shop/requirements.txt
Django==5.2
pillow==11.2.1
crispy-tailwing==1.0.3
djangorestframework==3.16.0